+++ element/alchemy_form.js
@@ -5,18 +5,7 @@
  * @since    0.1.0
  * @version  0.1.0
  */
-var Form = Function.inherits('Alchemy.Element.Form.Base', function Form() {
-	Form.super.call(this);
-});
-
-/**
- * The stylesheet to load for this element
- *
- * @author   Jelle De Loecker <jelle@develry.be>
- * @since    0.1.0
- * @version  0.1.0
- */
-Form.setStylesheetFile('form/form');
+var Form = Function.inherits('Alchemy.Element.Form.Base', 'Form');
 
 /**
  * The stylesheet to load for this element