+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.4 (WIP)
 
 * Make fields work with the new `Alchemy.Map.Backed` and `Alchemy.Map.Enum` class
+* Add the `field_path_in_current_schema` property
 
 ## 0.1.3 (2022-02-20)